requirements.txt
vlibrosa==0.9.2
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.1.3
tensorflow==2.9.1
matplotlib==3.6.2
